refactor: separate scene detection logic from extract_features_module (closes #14)
