+++ src/extract_features_module.py
@@ -4,10 +4,8 @@ import numpy as np
 import torchvision.models as models
 import torchvision.transforms as transforms
 import os
-import json
 from sklearn.decomposition import PCA
 from PIL import Image
-from transnetv2 import TransNetV2
 import platform 
 
 # OS에 따라 비디오 처리 라이브러리 선택
@@ -32,11 +30,11 @@ def extract_features_opencv(video_path, model, device, batch_size=32):
     if not cap.isOpened():
         raise IOError(f"비디오 파일을 열 수 없습니다: {video_path}")
 
-    fps = cap.get(cv2.CAP_PROP_FPS)
+    video_fps = cap.get(cv2.CAP_PROP_FPS)
     total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-    print(f"평균 FPS: {fps}")
+    print(f"평균 FPS: {video_fps}")
     
-    frame_idxs = list(range(0, total_frames, int(round(fps))))
+    frame_idxs = list(range(0, total_frames, int(round(video_fps))))
 
     transform = transforms.Compose([
         transforms.Resize((299, 299)),
@@ -67,16 +65,16 @@ def extract_features_opencv(video_path, model, device, batch_size=32):
             batch = []
     
     cap.release()
-    return np.concatenate(feats, axis=0)
+    return np.concatenate(feats, axis=0), video_fps
 
 # Decord를 사용한 특징 추출 함수 (Linux/Windows용)
 def extract_features_decord(video_path, model, device, batch_size=32):
     print("프레임 특징 추출 중... (Decord + 배치 처리, 메모리 최적화)")
     ctx = cpu(0)
     vr = VideoReader(video_path, ctx=ctx)
-    fps = vr.get_avg_fps()
-    print(f"평균 FPS: {fps}")
-    frame_idxs = list(range(0, len(vr), int(round(fps))))
+    video_fps = vr.get_avg_fps()
+    print(f"평균 FPS: {video_fps}")
+    frame_idxs = list(range(0, len(vr), int(round(video_fps))))
 
     transform = transforms.Compose([
         transforms.Resize((299, 299)),
@@ -101,7 +99,7 @@ def extract_features_decord(video_path, model, device, batch_size=32):
                 feats.append(batch_feats)
             batch = []
 
-    return np.concatenate(feats, axis=0)
+    return np.concatenate(feats, axis=0), video_fps
 
 # PCA 적용
 def apply_pca(features, max_components=1024):
@@ -122,34 +120,8 @@ def save_to_h5(features, output_h5):
     with h5py.File(output_h5, "w") as hf:
         hf.create_dataset("features", data=features)
 
-# TransNetV2를 이용한 장면 전환 감지
-def detect_scenes_transnetv2(video_path, threshold=0.5):
-    print("\n[2/6]TransNetV2로 장면 전환 감지 중...")
-    model = TransNetV2()
-    video_frames, single_frame_predictions, _ = model.predict_video(video_path)
-    scene_changes = np.where(single_frame_predictions > threshold)[0]
-    print(f"{len(scene_changes)}개의 장면 전환점 검출 완료")
-    return scene_changes.tolist(), video_frames.shape[0]
-
-# 장면 구간 JSON으로 저장
-def save_segments_to_json(scene_changes, output_json, total_frames, fps):
-    segment_data = []
-    scene_changes = [0] + scene_changes + [total_frames - 1]
-    for idx in range(len(scene_changes)-1):
-        start_frame = scene_changes[idx]
-        end_frame = scene_changes[idx+1]
-        segment_data.append({
-            "segment_id": idx,
-            "start_time": round(start_frame / fps, 2),
-            "end_time": round(end_frame / fps, 2)
-        })
-
-    with open(output_json, "w", encoding="utf-8") as f:
-        json.dump(segment_data, f, ensure_ascii=False, indent=4)
-    print("장면 구간 JSON 저장 완료")
-
 # 특징 추출 및 TransNetV2 장면 분할 파이프라인
-def extract_features_pipe(video_path, output_h5, output_json, device="cuda"):
+def extract_features_pipe(video_path, output_h5, device="cuda"):
     """
     비디오 파일에서 시각적 특징(feature)과 장면(scene) 정보를 추출하여 파일로 저장한다.
 
@@ -167,23 +139,17 @@ def extract_features_pipe(video_path, output_h5, output_json, device="cuda"):
     """
 
     os.makedirs(os.path.dirname(output_h5), exist_ok=True)
-    os.makedirs(os.path.dirname(output_json), exist_ok=True)
 
     model = load_inception_v3(device)
     
     # OS에 따라 다른 함수를 호출하여 특징 추출
     if IS_MACOS:
-        features = extract_features_opencv(video_path, model, device)
-        cap = cv2.VideoCapture(video_path)
-        fps = cap.get(cv2.CAP_PROP_FPS)
-        cap.release()
+        features, video_fps = extract_features_opencv(video_path, model, device)
     else:
-        features = extract_features_decord(video_path, model, device)
-        vr = VideoReader(video_path, ctx=cpu(0))
-        fps = vr.get_avg_fps()
-
+        features, video_fps = extract_features_decord(video_path, model, device)
+    
     pca_features = apply_pca(features)
     save_to_h5(pca_features, output_h5)
 
-    scene_changes, total_frames = detect_scenes_transnetv2(video_path)
-    save_segments_to_json(scene_changes, output_json, total_frames, fps)
+    # 비디오의 실제 초당 프레임 수
+    return video_fps
